requirements.txt
streamlit
numpy
matplotlib.pyplot